Move snowplow tracker to metabase/analytics/snowplow
